Trait pallet_executor::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type SecondaryHash: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + SimpleBitOps + Ord + Default + Copy + CheckEqual + Hash + AsRef<[u8]> + AsMut<[u8]> + MaybeMallocSizeOf + MaxEncodedLen;
    type ReceiptsPruningDepth: Get<Self::BlockNumber>;
    type MaximumReceiptDrift: Get<Self::BlockNumber>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type SecondaryHash: Parameter + Member + MaybeSerializeDeserialize + Debug + MaybeDisplay + SimpleBitOps + Ord + Default + Copy + CheckEqual + Hash + AsRef<[u8]> + AsMut<[u8]> + MaybeMallocSizeOf + MaxEncodedLen

Secondary chain block hash type.

type ReceiptsPruningDepth: Get<Self::BlockNumber>

Number of execution receipts kept in the state.

type MaximumReceiptDrift: Get<Self::BlockNumber>

Maximum execution receipt drift.

If the primary number of an execution receipt plus the maximum drift is bigger than the best execution chain number, this receipt will be rejected as being too far in the future.